requirements.txt
requests
pandas
psycopg2-binary
python-dotenv
matplotlib
streamlit
prophet
plotly
psycopg2-binary>=2.9.5
marshmallow<4
numpy
